Refactoring of My_Profile_TUI code to separate the Screens from the main App file and to rename the refactored App to my_profile_app.py
