+++ src/DemoCode/My_Profile/my_profile_app.py
@@ -1,1108 +1,25 @@
-"""
-   PDX-License-Identifier: Apache-2.0
-   Copyright Contributors to the ODPi Egeria project.
-
-   This file provides a set of report specification related functions for my_egeria.
-
-"""
-from __future__ import annotations
-
 import datetime
-import re
-from typing import Any
 
-from prompt_toolkit import PromptSession
-from prompt_toolkit.clipboard import ClipboardData
 from prompt_toolkit.clipboard.pyperclip import PyperclipClipboard
-
-from pyegeria.view.format_set_executor import exec_report_spec
-from pyegeria import MyProfile, PyegeriaException, print_basic_exception, AutomatedCuration
-from pyegeria import load_app_config, settings
-from pyegeria import MetadataExpert
-
-from textual.css.query import NoMatches
-
-from textual.widgets._tree import TreeNode
+from pyegeria import load_app_config, settings, MyProfile, PyegeriaException, print_basic_exception, exec_report_spec, \
+    AutomatedCuration, MetadataExpert
 from textual import on, work
 from textual.app import App, ComposeResult
-from textual.containers import ScrollableContainer, Container
-from textual.screen import ModalScreen, Screen
-from textual.widgets import (
-    Button,
-    DataTable,
-    Footer,
-    Header,
-    Input,
-    Static, Tree, OptionList, TextArea, Placeholder,
-)
+from textual.containers import ScrollableContainer
+from textual.widgets import DataTable, OptionList, Header, Static, Footer, Tree
 from textual.widgets._option_list import Option
 
-class SelectionOverviewScreen(Screen):
-    """Screen to display the selection of the user's data sources."""
-    BINDINGS = [("q", "quit", "Quit"),
-                ("b", "back", "Go back")]
-
-    CSS_PATH = "my_profile.tcss"
-
-    def __init__(self, category, view_server, url, user, pwd):
-        self.category = category
-        self.view_server = view_server
-        self.platform_url = url
-        self.user_name = user
-        self.user_password = pwd
-        if self.category == "glossary":
-            self.data_tree: Tree = app.query_one("#glossary_details_tree", Tree)
-        elif self.category == "catalog":
-            self.data_tree: Tree = app.query_one("#digital_product_catalog_tree", Tree)
-        elif self.category == "dictionary":
-            self.data_tree: Tree = app.query_one("#data_dictionary_tree", Tree)
-        elif self.category == "domain":
-            self.data_tree: Tree = app.query_one("#business_domain_tree", Tree)
-        elif self.category == "specification":
-            self.data_tree: Tree = app.query_one("#data_specification_tree", Tree)
-        else:
-            # unknown category
-            # push status screen, logic error in code.
-            error_category = "Collection Category"
-            error_message = "Unknown collection category returned"
-            self.log(f"Error in selection overview processing: {error_category}, {error_message}")
-            app.push_screen(StatusScreen(f"{error_category}: {error_message}"), callback=app.status_callback)
-        super().__init__()
-
-    def compose(self) -> ComposeResult:
-        """ Compose the UI components for the SelectionOverviewScreen screen."""
-        self.title = f"Shopping for Data, Data Selection:"
-        self.sub_title = f"Category: {self.category}"
-        yield Header(show_clock=True)
-        yield Static("Please select an item from the tree [blink]:[/]")
-        yield ScrollableContainer(
-                                    self.data_tree,
-                                    id="data_tree_container"
-                                    )
-        yield ScrollableContainer(
-                                Placeholder(id="data_details_placeholder"),
-                                id="data_details_placeholder_container"
-                                    )
-        yield Footer()
-
-    def action_quit(self) -> None:
-        """ The quit option in the footer has been selected. Dismiss the screen."""
-        self.dismiss(210)
-
-    def action_back(self) -> None:
-        """ The back option in the footer has been selected. Dismiss the screen."""
-        self.dismiss(200)
-
-    @on(Tree.NodeSelected, "#data_tree")
-    def handle_tree_node_selected(self, event: Tree.NodeSelected):
-        # Modify to handle for each tree?
-        self.node_selected = event.node
-        self.log(f"Node selected: {self.node_selected}")
-        self.node_label = self.node_selected.label
-        # the provided id can be either a GUID or a qualified name!
-        # the variable is labeled guid but it could contain a qualified name, both guid and qualified name are strings.
-        self.node_GUID = self.node_selected.data
-        self.log(f"Node label: {self.node_label}, GUID: {self.node_GUID}")
-        if self.node_label == "glossary":
-            self.display_selected_term_details(self.node_GUID)
-        elif self.node_label == "digital_product_catalog":
-            self.display_selected_digital_product(self.node_GUID)
-        elif self.node_label == "data_dictionary":
-            self.display_selected_data_dictionary(self.node_GUID)
-        elif self.node_label == "business_domain":
-            self.display_selected_business_domain(self.node_GUID)
-        elif self.node_label == "data_specification":
-            self.display_selected_data_specification(self.node_GUID)
-
-    def display_selected_term_details(self, term_GUID) -> Any:
-        """ The user has selected a glossary term, build a display of the term details,
-             and show along side the glossary tree """
-        self.term_GUID = term_GUID
-        self.log(f"Selected glossary term GUID: {self.term_GUID}")
-        try:
-            self.glossary_term_data = exec_report_spec(format_set_name="Glossary-Terms",
-                                                  output_format="DICT",
-                                                  params={"search_string": self.term_GUID, "filter_string": self.term_GUID},
-                                                  view_server=self.view_server,
-                                                  view_url=self.platform_url,
-                                                  user=self.user_name,
-                                                  user_pass=self.user_password)
-        except PyegeriaException as e:
-            print_basic_exception(e)
-            self.log(f"Error retrieving glossary details: {e!s}")
-            self.dismiss(421)
-            return (421)
-        if not self.glossary_term_data or self.glossary_term_data == []:
-            self.log(f"No glossary term data returned for GUID: {self.term_GUID}")
-            Static(f"No glossary term data returned for GUID: {self.term_GUID}").mount(self.query_one("#glossary_term_details_container"))
-        else:
-            self.log(f"Glossary term data returned for GUID: {self.term_GUID}")
-            TextArea(f"Glossary term data returned for GUID: {self.term_GUID}", id="glossary_term_details_text_area", read_only=True).mount(self.query_one("#glossary_term_details_container"))
-            TextArea.data = self.glossary_term_data
-
-    def display_selected_digital_product(self, digital_product_GUID) -> Any:
-        """ The user has selected a glossary term, build a display of the term details,
-                    and show along side the glossary tree """
-        self.digital_product_GUID = digital_product_GUID
-        self.log(f"Selected digital product GUID: {self.digital_product_GUID}")
-        try:
-            self.digital_product_data = exec_report_spec(format_set_name="Digital-Products",
-                                                       output_format="DICT",
-                                                       params={"search_string": self.digital_product_GUID,
-                                                               "filter_string": self.digital_product_GUID},
-                                                       view_server=self.view_server,
-                                                       view_url=self.platform_url,
-                                                       user=self.user_name,
-                                                       user_pass=self.user_password)
-        except PyegeriaException as e:
-            print_basic_exception(e)
-            self.log(f"Error retrieving digital product details: {e!s}")
-            self.dismiss(421)
-            return (421)
-        if not self.digital_product_data or self.digital_product_data == []:
-            self.log(f"No digital product data returned for GUID: {self.digital_product_GUID}")
-            Static(f"No digital product data returned for GUID: {self.digital_product_GUID}").mount(
-                self.query_one("#digital_product_details_container"))
-        else:
-            self.log(f"Digital product data returned for GUID: {self.digital_product_GUID}")
-            TextArea(f"Digital product data returned for GUID: {self.digital_product_GUID}", id="digital_product_details_text_area",
-                     read_only=True).mount(self.query_one("#digital_product_details_container"))
-            TextArea.data = self.digital_product_data
-
-    def display_selected_data_dictionary(self, data_dictionary_GUID) -> Any:
-        """ The user has selected a data dictionary, build a display of the dictionary details,
-                            and show along side the dictionary tree """
-        self.data_dictionary_GUID = data_dictionary_GUID
-        self.log(f"Selected data dictionary GUID: {self.data_dictionary_GUID}")
-        try:
-            self.data_dictionary_data = exec_report_spec(format_set_name="Data-Dictionaries",
-                                                         output_format="DICT",
-                                                         params={"search_string": self.data_dictionary_GUID,
-                                                                 "filter_string": self.data_dictionary_GUID},
-                                                         view_server=self.view_server,
-                                                         view_url=self.platform_url,
-                                                         user=self.user_name,
-                                                         user_pass=self.user_password)
-        except PyegeriaException as e:
-            print_basic_exception(e)
-            self.log(f"Error retrieving data dictionary details: {e!s}")
-            self.dismiss(421)
-            return (421)
-        if not self.data_dictionary_data or self.data_dictionary_data == []:
-            self.log(f"No data dictionary data returned for GUID: {self.data_dictionary_GUID}")
-            Static(f"No data dictionary data returned for GUID: {self.data_dictionary_GUID}").mount(
-                self.query_one("#data_dictionary_details_container"))
-        else:
-            self.log(f"Data dictionary data returned for GUID: {self.data_dictionary_GUID}")
-            TextArea(f"Data dictionary data returned for GUID: {self.data_dictionary_GUID}",
-                     id="data_dictionary_details_text_area",
-                     read_only=True).mount(self.query_one("#data_dictionary_details_container"))
-            TextArea.data = self.data_dictionary_data
-
-    def display_selected_business_domain(self, business_domain_GUID) -> int:
-        """ The user has selected a business domain, build a display of the domain details,
-                            and show along side the glossary tree """
-        self.business_domain_GUID = business_domain_GUID
-        self.log(f"Selected business domain GUID: {self.business_domain_GUID}")
-        try:
-            self.business_domain_data = exec_report_spec(format_set_name="BusinessCapabilities",
-                                                         output_format="DICT",
-                                                         params={"search_string": self.business_domain_GUID,
-                                                                 "filter_string": self.business_domain_GUID},
-                                                         view_server=self.view_server,
-                                                         view_url=self.platform_url,
-                                                         user=self.user_name,
-                                                         user_pass=self.user_password)
-        except PyegeriaException as e:
-            print_basic_exception(e)
-            self.log(f"Error retrieving business domain details: {e!s}")
-            self.dismiss(421)
-            return (421)
-        if not self.business_domain_data or self.business_domain_data == []:
-            self.log(f"No business domain data returned for GUID: {self.business_domain_GUID}")
-            Static(f"No business domain data returned for GUID: {self.business_domain_GUID}").mount(
-                self.query_one("#business_domain_details_container"))
-        else:
-            self.log(f"Business domain data returned for GUID: {self.business_domain_GUID}")
-            TextArea(f"Business domain data returned for GUID: {self.business_domain_GUID}",
-                     id="business_domain_details_text_area",
-                     read_only=True).mount(self.query_one("#business_domain_details_container"))
-            TextArea.data = self.business_domain_data
-
-    def display_selected_data_specification(self, data_specification_GUID) -> int:
-        """ The user has selected a data specification, build a display of the specification details
-            and show alongside data specifications tree """
-
-        self.data_specification_qualified_name = data_specification_GUID
-        self.log(f"Data specification selected: {self.data_specification_qualified_name}")
-        try:
-            self_data_specification_data = exec_report_spec(format_set_name="Data-Specification",
-                                                            output_format="DICT",
-                                                            params={"search_string": self.data_specification_qualified_name,
-                                                                    "filter_string": self.data_specification_qualified_name},
-                                                            view_server=self.view_server,
-                                                            view_url=self.platform_url,
-                                                            user=self.user_name,
-                                                            user_pass=self.user_password)
-        except PyegeriaException as e:
-            print_basic_exception(e)
-            self.log(f"Error retrieving data specification details: {e!s}")
-            self.dismiss(431)
-            return (431)
-        if not self_data_specification_data or self_data_specification_data == []:
-            self.log(f"No data specification data returned for qualified name: {self.data_specification_qualified_name}")
-            Static(f"No data specification data returned for qualified name: {self.data_specification_qualified_name}").mount(
-                self.query_one("#data_specification_details_container"))
-        else:
-            self.log(f"Data specification data returned for qualified name: {self.data_specification_qualified_name}")
-            TextArea(f"Data specification data returned for qualified name: {self.data_specification_qualified_name}",
-                     id="data_specification_details_text_area",
-                     read_only=True).mount(self.query_one("#data_specification_details_container"))
-            TextArea.data = self_data_specification_data
-
-
-class ShopForDataScreen(Screen):
-    """ Screen to Present a choice of different data sources to the user."""
-    BINDINGS = [("q", "dismiss(200)", "Quit"),
-                ("b", "back", "Go back")]
-
-    CSS_PATH = "my_profile.tcss"
-
-    def __init__ (self, glossary_table, digital_product_catalog_table, data_dictionary_table, business_domain_table, data_specification_table ):
-        """Initialize the ShopForDataScreen screen."""
-        self.glossary_table: DataTable = glossary_table
-        self.digital_product_catalog_table: DataTable = digital_product_catalog_table
-        self.data_dictionary_table: DataTable = data_dictionary_table
-        self.business_domain_table: DataTable = business_domain_table
-        self.data_specification_table: DataTable = data_specification_table
-        super().__init__()
-
-    def compose(self) -> ComposeResult:
-        """ Compose the UI components for the ShopForDataScreen screen."""
-        yield Header(show_clock=True)
-        yield TextArea(f"\n\n\nPlease click on a line to select a data source:")
-        yield ScrollableContainer(
-            Static("Glossary"),
-            self.glossary_table)
-        yield ScrollableContainer(
-            Static("Digital Product Catalog"),
-            self.digital_product_catalog_table)
-        yield ScrollableContainer(
-            Static("Data Dictionary"),
-            self.data_dictionary_table)
-        yield ScrollableContainer(
-            Static("Business Domain"),
-            self.business_domain_table)
-        yield ScrollableContainer(
-            Static("Data Specification"),
-            self.data_specification_table)
-        yield Footer()
-
-    @on(DataTable.RowSelected, "#glossary_table")
-    def handle_glossary_table_selection(self, event: DataTable.RowSelected):
-        row_selected = event.row_key
-        row_values = self.glossary_table.get_row(row_selected)
-        row_display_name = row_values[0]
-        row_description = row_values[1]
-        row_qualified_name = row_values[2]
-        self.log(f"Row selected: {row_selected}, values: {row_values}, display name: {row_display_name}, description: {row_description}, qualified name: {row_qualified_name}")
-        self.dismiss (["glossary", row_qualified_name, row_display_name])
-
-    @on(DataTable.RowSelected, "#digital_product_catalog_table")
-    def handle_digital_product_catalog_table_selection(self, event: DataTable.RowSelected):
-        row_selected = event.row_key
-        self.log(f"Digital product row selected: {row_selected}, event: {event}")
-        # row_values = self.glossary_table.get_row(event.row_key)
-        row_values = self.query_one("#digital_product_catalog_table", DataTable).get_row(event.row_key)
-        self.log(f"Digital product row values: {row_values}")
-        row_display_name = row_values[0]
-        row_description = row_values[1]
-        row_qualified_name = row_values[2]
-        self.log(f"Row selected: {row_selected}, values: {row_values}, display name: {row_display_name}, description: {row_description}, qualified name: {row_qualified_name}")
-        self.dismiss (["product", row_qualified_name, row_display_name])
-
-    @on(DataTable.RowSelected, "#data_dictionary_table")
-    def handle_data_dictionary_table_selection(self, event: DataTable.RowSelected):
-        row_selected = event.row_key
-        row_values = self.query_one("#data_dictionary_table", DataTable).get_row(event.row_key)
-        row_display_name = row_values[0]
-        row_description = row_values[1]
-        row_qualified_name = row_values[2]
-        self.log(f"Row selected: {row_selected}, values: {row_values}, display name: {row_display_name}, description: {row_description}, qualified name: {row_qualified_name}")
-        self.dismiss (["dictionary", row_qualified_name, row_display_name])
-
-    @on(DataTable.RowSelected, "#business_domain_table")
-    def handle_business_domain_table_selection(self, event: DataTable.RowSelected):
-        row_selected = event.row_key
-        row_values = self.query_one("#business_domain_table", DataTable).get_row(event.row_key)
-        row_qualified_name = row_values[0]
-        row_type_name = row_values[1]
-        row_guid = row_values[2]
-        self.log(f"Row selected: {row_selected}, values: {row_values}, qualified name: {row_qualified_name}, type name: {row_type_name}, guid: {row_guid}")
-        self.dismiss (["domain", row_qualified_name, row_type_name])
-
-    @on(DataTable.RowSelected, "#data_specification_table")
-    def handle_data_specification_table_selection(self, event: DataTable.RowSelected):
-        row_selected = event.row_key
-        row_values = self.query_one("#data_specification_table", DataTable).get_row(event.row_key)
-        row_display_name = row_values[0]
-        row_description = row_values[1]
-        row_qualified_name = row_values[2]
-        self.log(f"Row selected: {row_selected}, values: {row_values}, display name: {row_display_name}, description: {row_description}, qualified name: {row_qualified_name}")
-        self.dismiss (["specification", row_qualified_name, row_display_name])
-
-    def action_back(self) -> None:
-        """ The back option in the footer has been selected. Dismiss the screen."""
-        self.dismiss(200)
-
-    def action_quit(self) -> None:
-        """ The quit option in the footer has been selected. Dismiss the screen."""
-        self.dismiss(210)
-
-
-class StatusScreen(ModalScreen[Any]):
-    """Modal screen to display the status of the application."""
-    BINDINGS = [("q", "quit", "Quit"),
-                ("return", "successful", "Continue"),
-                ("b", "unsuccessful", "Bad result"),
-                ("c", "copy GUID to clipboard", "Copy GUID to clipboard"),
-                ]
-
-    CSS_PATH = "my_profile.tcss"
-
-    def __init__(self, status_message: str) -> None:
-        """Initialize the StatusScreen screen."""
-        super().__init__()
-        self.status_message = status_message
-        self.clipboard = PyperclipClipboard()
-
-    def compose(self) -> ComposeResult:
-        """ Compose the UI components for the StatusScreen screen."""
-        yield Header(show_clock=True)
-        yield TextArea(self.status_message, id="status_message_text_area", read_only=True)
-        yield Footer()
-        
-    def action_quit(self) -> None:
-        """ The quit option in the footer has been selected. Dismiss the screen."""
-        self.dismiss(200)
-        
-    def action_unsuccessful(self) -> None:
-        """ The bad result option in the footer has been selected. Dismiss the screen."""
-        self.dismiss(400)
-        
-    def action_copy_guid_to_clipboard(self) -> None:
-        """ The copy GUID to clipboard option in the footer has been selected. Copy the GUID to the clipboard."""
-        # The GUID is between single ' marks in the status_message when a good result is recorded.
-        match = re.search(r"'(.*?)'", self.status_message)
-        self.log(f"Match: {match}")
-        if match:
-            guid = match.group(1)
-            self.clipboard.set_data(ClipboardData(text = guid))
-            self.log(f"Copied GUID to clipboard: {guid}")
-        else:
-            self.dismiss(400)
-        self.dismiss(200)
-
-    def action_successful(self):
-        """ The successful option in the footer has been selected. Dismiss the screen."""
-        self.dismiss(200)
-
-class TechnologyTypeProcessesScreen(ModalScreen[Any]):
-    """Modal screen to display technology type templates in Egeria."""
-    BINDINGS = [("q", "quit", "Quit"),
-                ("b", "back", "Go back"),
-                ("ctl+e", "expand", "Toggle Twisties")]
-
-    CSS_PATH = "my_profile.tcss"
-
-    def __init__ (self,
-                  user_name: str,
-                  user_kpts: int,
-                  tech_type_name: str,
-                  tech_type_description: str,
-                  selected_t_option,
-                  tech_type_option_selected,
-                  tech_type_processes
-                  ) -> None:
-        """Initialize the Technology_Type_Templatess screen."""
-        super().__init__()
-        self.user_name = user_name
-        self.karma_points = user_kpts
-        self.tech_type_name = tech_type_name
-        self.tech_type_description = tech_type_description
-        self.selected_t_option = selected_t_option
-        self.selected_t_option_selected = tech_type_option_selected
-        self.tech_type_processes = tech_type_processes
-        self.full_process = None
-        load_app_config()
-
-    async def on_mount(self) -> None:
-        """ On Mount function of the Technology_Type_Templatess screen."""
-        self.title = f"User: {self.user_name}, Karma Points: {self.karma_points}"
-        self.sub_title = f"Technology Type: {self.tech_type_name}, Description: {self.tech_type_description}"
-
-        if self.selected_t_option == "process":
-            self.log(f"Processing processes, with data: {self.selected_t_option_selected}")
-            # get selected process from the tech_type data
-            self.log(f"Technology Type Process: {self.tech_type_processes}")
-            if isinstance(self.tech_type_processes, list):
-                for process in self.tech_type_processes:
-                    if process.get("Governance Process Name") == self.selected_t_option_selected:
-                        self.full_process = process
-                        self.selected_t_process = process.get("Qualified Name")
-                        self.log(f"Selected Process: {self.selected_t_process}")
-                        break
-                    else:
-                        continue
-            self.log(f"Selected Process: {self.selected_t_process}")
-
-            for placeholder in self.selected_t_process:
-                name = placeholder.get("Property Name") or None
-                Description = placeholder.get("Description") or None
-                Type = placeholder.get("Data Type") or None
-                Example = placeholder.get("Example") or None
-                Required = placeholder.get("Required") or False
-
-                # Sanitize the name for use as a CSS ID
-                safe_name = name.replace(" ", "_") if name else f"placeholder_{id(placeholder)}"
-                placeholder_text: TextArea = TextArea(
-                    f"{name}\n\nDescription: {Description}\nType: {Type}\nExample: {Example}\nRequired: {Required}",
-                    id=f"{safe_name}_placeholder_text_area",
-                    read_only=True
-                )
-                # Ensure TextArea is visible
-                placeholder_text.styles.height = 8
-
-                placeholder_input = Input(id=f"{safe_name}_placeholder_input", placeholder="Enter value here")
-                self.log(f"Placeholder: {placeholder_text.text}\n {placeholder_input}")
-
-                # Mount the TextArea and the associated Input field into the ScrollableContainer
-                try:
-                    load_point = self.query_one("#technology_type_templates_input")
-                    await load_point.mount(placeholder_text, before="#submit_button")
-                    await load_point.mount(placeholder_input, before="#submit_button")
-                    self.log(f"Placeholder text area loaded: {placeholder_text.text}")
-                    self.log(f"Placeholder input loaded: {placeholder_input}")
-                    continue
-                except Exception as e:
-                    self.log(f"Error loading placeholder container: {e!s}")
-                    self.app.dismiss(416)
-
-    def compose(self) -> ComposeResult:
-        """ Compose the UI components for the Technology_Type_Templatess screen."""
-        yield Header(show_clock=True)
-        yield Static("Please complete the required fields and any optional fields you prefer:")
-        yield ScrollableContainer(
-            Static("Technology Type Template Input"),
-            Button("Submit", id="submit_button"),
-            id="technology_type_templates_input"
-        )
-        yield Footer()
-
-    def action_quit(self) -> None:
-        """ The quit option in the footer has been selected. Dismiss the screen."""
-        self.dismiss("200", )
-
-    @on(Button.Pressed, "#submit_button")
-    def handle_submit_button_pressed(self, event: Button.Pressed) -> None:
-        """ The submit button has been pressed."""
-        self.log(f"Submit button pressed, button: {event.button}")
-        save_input_data: dict = {}
-        for input_widget in self.query("Input"):
-            self.log(f"Input widget: {input_widget.id}, value: {input_widget}")
-            save_input_data.update({input_widget.id: input_widget})
-        self.log(f"Save input data: {save_input_data}")
-        self.dismiss(["input", save_input_data, self.full_process])
-
-    @on(Input.Changed, "#technology_type_templates_input")
-    def handle_input_changed(self, event: Input.Changed) -> None:
-        """ The user has changed the input on the screen."""
-        self.log(f"Input changed, input: {event.input}")
-
-
-class TechnologyTypeTemplatesScreen(ModalScreen[Any]):
-    """Modal screen to display technology type templates in Egeria."""
-    BINDINGS = [("q", "dismiss(200)", "Quit"),
-                ("b", "back", "Go back"),
-                ("ctl+e", "expand", "Toggle Twisties")]
-
-    CSS_PATH = "my_profile.tcss"
-
-    def __init__ (self,
-                  user_name: str,
-                  user_kpts: int,
-                  tech_type_name: str,
-                  tech_type_description: str,
-                  selected_t_option,
-                  tech_type_option_selected,
-                  tech_type_templates
-                  ) -> None:
-        """Initialize the Technology_Type_Templatess screen."""
-        super().__init__()
-        self.user_name = user_name
-        self.karma_points = user_kpts
-        self.tech_type_name = tech_type_name
-        self.tech_type_description = tech_type_description
-        self.selected_t_option = selected_t_option
-        self.selected_t_option_selected = tech_type_option_selected
-        self.tech_type_templates = tech_type_templates
-        self.full_template = None
-        load_app_config()
-
-    async def on_mount(self) -> None:
-        """ On Mount function of the Technology_Type_Templatess screen."""
-        self.title = f"User: {self.user_name}, Karma Points: {self.karma_points}"
-        self.sub_title = f"Technology Type: {self.tech_type_name}, Description: {self.tech_type_description}"
-
-        if self.selected_t_option == "template":
-            self.log(f"Processing templates, with data: {self.selected_t_option_selected}")
-            # get selected template from the tech_type data
-            self.log(f"Technology Type Templates: {self.tech_type_templates}")
-            if isinstance(self.tech_type_templates, list):
-                for template in self.tech_type_templates:
-                    if template.get("Catalog Template Name") == self.selected_t_option_selected:
-                        self.full_template = template
-                        self.selected_t_template = template.get("Placeholder Properties")
-                        self.log(f"Selected Template: {self.selected_t_template}")
-                        break
-                    else:
-                        continue
-            self.log(f"Selected Template: {self.selected_t_template}")
-
-            for placeholder in self.selected_t_template:
-                name = placeholder.get("Property Name") or None
-                Description = placeholder.get("Description") or None
-                Type = placeholder.get("Data Type") or None
-                Example = placeholder.get("Example") or None
-                Required = placeholder.get("Required") or False
-
-                # Sanitize the name for use as a CSS ID
-                safe_name = name.replace(" ", "_") if name else f"placeholder_{id(placeholder)}"
-                placeholder_text: TextArea = TextArea(
-                    f"{name}\n\nDescription: {Description}\nType: {Type}\nExample: {Example}\nRequired: {Required}",
-                    id=f"{safe_name}_placeholder_text_area",
-                    read_only=True
-                )
-                # Ensure TextArea is visible
-                placeholder_text.styles.height = 8
-
-                placeholder_input = Input(id=f"{safe_name}_placeholder_input", placeholder="Enter value here")
-                self.log(f"Placeholder: {placeholder_text.text}\n {placeholder_input}")
-
-                # Mount the TextArea and the associateds Input field into the ScrollableContainer
-                try:
-                    load_point = self.query_one("#technology_type_templates_input")
-                    await load_point.mount(placeholder_text, before="#submit_button")
-                    await load_point.mount(placeholder_input, before="#submit_button")
-                    self.log(f"Placeholder text area loaded: {placeholder_text.text}")
-                    self.log(f"Placeholder input loaded: {placeholder_input}")
-                    continue
-                except Exception as e:
-                    self.log(f"Error loading placeholder container: {e!s}")
-                    self.app.dismiss(416)
-
-    def compose(self) -> ComposeResult:
-        """ Compose the UI components for the Technology_Type_Templatess screen."""
-        yield Header(show_clock=True)
-        yield Static("Please complete the required fields and any optional fields you prefer:")
-        yield ScrollableContainer(
-            Static("Technology Type Template Input"),
-            Button("Submit", id="submit_button"),
-            id="technology_type_templates_input"
-        )
-        yield Footer()
-
-    def action_quit(self) -> None:
-        """ The quit option in the footer has been selected. Dismiss the screen."""
-        self.dismiss("200", )
-
-    @on(Button.Pressed, "#submit_button")
-    def handle_submit_button_pressed(self, event: Button.Pressed) -> None:
-        """ The submit button has been pressed."""
-        self.log(f"Submit button pressed, button: {event.button}")
-        save_input_data: dict = {}
-        for input_widget in self.query("Input"):
-            self.log(f"Input widget: {input_widget.id}, value: {input_widget}")
-            save_input_data.update({input_widget.id: input_widget})
-        self.log(f"Save input data: {save_input_data}")
-        self.dismiss(["input", save_input_data, self.full_template])
-
-    @on(Input.Changed, "#technology_type_templates_input")
-    def handle_input_changed(self, event: Input.Changed) -> None:
-        """ The user has changed the input on the screen."""
-        self.log(f"Input changed, input: {event.input}")
-
-
-class TechnologyTypeOptionsScreen(ModalScreen):
-    """ Modal screen to display a technology type's templates and processes."""
-    BINDINGS = [("q", "dismiss", "Quit"),
-                ("b", "back", "Go back"),
-                ("ctl+s", "Select", "tech_type_option_select")]
-
-    CSS_PATH = "my_profile.tcss"
-
-    def __init__(self, tech_type_guid: str,
-                       tech_type_name: str,
-                       tech_type_description: str,
-                       user_name: str,
-                       user_pwd:str,
-                       user_kpts: int,
-                       tech_type_templates: list[dict],
-                       tech_type_processes: list[dict]) -> None:
-        """Initialize the TechnologyTypeOptions screen with a technology type's templates and processes."""
-        super().__init__()
-        self.selected_process_index = None
-        self.tech_type_guid = tech_type_guid
-        self.tech_type_name = tech_type_name
-        self.tech_type_description = tech_type_description
-        self.user_name = user_name
-        self.user_password = user_pwd
-        self.karma_points = user_kpts
-        self.tech_type_templates = tech_type_templates
-        self.tech_type_processes = tech_type_processes
-        self.selected_template_guid = None
-        self.selected_process_guid = None
-        self.option_type_selected = None
-        self.selected_template = None
-        self.selected_template_index = None
-        self.selected_template_data = None
-        self.selected_process = None
-        self.selected_process_data = None
-
-        load_app_config("config/config.json")
-        app_config = settings.Environment
-        app_user = settings.User_Profile
-
-        if not self.user_name:
-            self.user_name = app_user.user_name or "garygeeke"
-            self.user_password = app_user.user_pwd or "secret"
-
-    def compose(self) -> ComposeResult:
-
-        """ Compose the UI components for the TechnologyTypeOptions screen."""
-        yield Header(show_clock=True)
-        yield Static(f"Description: {self.tech_type_description}")
-        yield ScrollableContainer(
-            Static(f"Technology Type: {self.tech_type_name}"),
-            Container(
-                Static("Select a template or process to continue."),
-                Static(f"Templates", id="number_of_templates_label"),
-                ScrollableContainer(
-                    Static("Available Templates", id="template_options_label"),
-                    OptionList(id="template_options"),
-                    Button("Select Template", id="select_template_btn"),
-                    id="template_options_container"),
-                Static(f"Processes:", id="number_of_processes_label"),
-                ScrollableContainer(
-                    Static("Available Processes",id="process_options_label"),
-                    OptionList(id="process_options"),
-                    id="process_options_container"),
-                Button(
-                    "Select Process",
-                    id="select_process_btn")))
-        yield Footer()
-
-    async def on_mount(self) -> int:
-        """Mount the TechnologyTypeOptions screen."""
-        self.title =  f"User: {self.user_name}, Karma Points: {self.karma_points}"
-        self.sub_title = f"Technology Type: {self.tech_type_name}, Description: {self.tech_type_description}"
-        self.log(f"Technology Type: {self.tech_type_name}, Description: {self.tech_type_description}")
-
-        for widget in self.query():
-            self.log(f"Widgets: {widget}")
-
-        self.log(f"Templates: {self.tech_type_templates}, Processes: {self.tech_type_processes}")
-        self.log(f"Templates Type: {type(self.tech_type_templates)}, Processes Type: {type(self.tech_type_processes)}")
-
-        if self.tech_type_templates and self.tech_type_templates != "None":
-            for t in self.tech_type_templates:
-                try:
-                    self.log(f"Template: {t}")
-                    templates = self.query_one("#template_options", OptionList).add_option(
-                        Option(t.get("Catalog Template Name")))
-                    self.log(f"Added option: {t.get('Catalog Template Name')}")
-                    await templates.mount(after=self.query_one("#template_options_label"))
-                    self.log(f"Mounted option")
-                    self.query_one("#select_template_btn", Button).disabled = False
-                    self.log(f"Enabled the template button")
-                except NoMatches as e:
-                    try:
-                        templates: OptionList = OptionList(id="template_options")
-                        await templates.mount(after=self.query_one("#template_options_label"))
-                        templates.add_option(Option(t.get("Catalog Template Name")))
-                        templates.refresh()
-                        self.query_one("#select_template_btn", Button).disabled = False
-                        continue
-                    except Exception as e:
-                        self.log(f"Error creating template option list: {e}, (410")
-                        return (410)
-                except Exception as e:
-                    self.log(f"Error creating template option list: {e} (411)")
-                    return (411)
-        else:
-            try:
-                self.log(f" No Templates")
-                templates = self.query_one("#template_options", OptionList).add_option(
-                    Option("No Templates Found for this Tech Type"))
-                await templates.mount(after=self.query_one("#template_options_label"))
-                self.query_one("#select_template_btn", Button).disabled = True
-            except NoMatches as e:
-                try:
-                    templates: OptionList = OptionList(id="template_options")
-                    await templates.mount(after=self.query_one("#template_options_label"))
-                    templates.add_option(Option("No Templates found for this Tech Type"))
-                    self.query_one("#select_template_btn", Button).disabled = True
-                except Exception as e:
-                    self.log(f"Error creating template option list: {e}")
-                    return (409)
-            except Exception as e:
-                self.log(f"Error creating template option list: {e}")
-                return (408)
-
-        if self.tech_type_processes and self.tech_type_processes != "None":
-            for p in self.tech_type_processes:
-                try:
-                    self.log(f"Process: {p}")
-                    processes = self.query_one("#process_options", OptionList).add_option(
-                        Option(p.get("Governance Process Name")))
-                    await processes.mount(after=self.query_one("#process_options_label"))
-                    self.query_one("#select_process_btn", Button).disabled = False
-                except NoMatches as e:
-                    try:
-                        processes: OptionList = OptionList(id="process_options")
-                        await processes.mount(after=self.query_one("#process_options_label"))
-                        processes.add_option(Option(p.get("Governance Process Name")))
-                        processes.refresh()
-                        self.query_one("#select_process_btn", Button).disabled = False
-                    except Exception as e:
-                        self.log(f"Error creating process option list: {e}")
-                        return (407)
-                except Exception as e:
-                    self.log(f"Error creating process option list: {e}")
-                    return (406)
-        else:
-            try:
-                self.log(f" No Processes")
-                processes = self.query_one("#process_options", OptionList).add_option(
-                    Option("No processes found for this Tech Type"))
-                await processes.mount(after=self.query_one("#process_options_label"))
-                self.query_one("#select_process_btn", Button).disabled = True
-            except NoMatches as e:
-                try:
-                    processes: OptionList = OptionList(id="process_options")
-                    await processes.mount(after=self.query_one("#process_options_label"))
-                    processes.add_option(Option(" No Processes found for this Tech Type"))
-                    processes.refresh()
-                    self.query_one("#select_process_btn", Button).disabled = True
-                except Exception as e:
-                    self.log(f"Error creating process option list: {e}")
-                    return (405)
-            except Exception as e:
-                self.log(f"Error creating process option list: {e}")
-                return (404)
-
-        return(200)
-
-    @on(OptionList.OptionHighlighted, "#template_options")
-    def handle_template_option_highlighted(self, event: OptionList.OptionHighlighted) -> None:
-        """Handle the highlighting of an option in the template option list."""
-        self.log(f"Template option highlighted: {event.option}")
-        self.log(f"Option index: {event.option_index}, Option list: {event.option_list.name}")
-        self.selected_template_index = event.option_index
-        selected_option = event.option
-        self.selected_template = selected_option.prompt
-        self.log(f"Highlighted Template: {self.selected_template}, index: {self.selected_template_index}")
-
-    @on(OptionList.OptionSelected, "#template_options")
-    def handle_template_option_selected(self, event: OptionList.OptionSelected) -> None:
-        """Handle the selection of an option from the template or process option lists."""
-        self.log(f"Template selected: {event.option}")
-        self.log(f"Option index: {event.option_index}, Option list: {event.option_list.name}")
-        self.option_type_selected = event.option_list.name
-        self.selected_template_index = event.option_index
-        selected_option = event.option
-        self.selected_template = selected_option.prompt
-        self.log(f"Selected Option List: {self.option_type_selected}, template: {self.selected_template}, index: {self.selected_template_index}")
-
-    @on(OptionList.OptionHighlighted, "#process_options")
-    def handle_process_option_highlighted(self, event: OptionList.OptionHighlighted) -> None:
-        """Handle the highlighting of an option in the process option list."""
-        self.log(f"Process option highlighted: {event.option}")
-        self.log(f"Option index: {event.option_index}, Option list: {event.option_list.name}")
-        self.selected_process_index = event.option_index
-        selected_option = event.option
-        self.selected_process = selected_option.prompt
-        self.log(f"Highlighted Process: {self.selected_process}, index: {self.selected_process_index}")
-
-    @on(OptionList.OptionSelected, "#process_options")
-    def handle_process_option_selected(self, event: OptionList.OptionSelected) -> None:
-        """Handle the selection of an option from the template or process option lists."""
-        self.log(f"Process selected: {event.option}")
-        self.log(f"Option index: {event.option_index}, Option list: {event.option_list.name}")
-        self.option_type_selected = event.option_list.name
-        self.selected_process_index = event.option_index
-        selected_option = event.option
-        self.selected_process = selected_option.prompt
-        self.log(f"Selected Option List: {self.option_type_selected}, template: {self.selected_process}, index: {self.selected_process_index}")
-
-    def action_back(self) -> None:
-        """Handle the back button press."""
-        self.dismiss("back")
-
-    @on(Button.Pressed, "#select_template_btn")
-    def handle_template_selected(self, event: Button.Pressed) -> None:
-        """Handle the selection of a template option."""
-        if self.selected_template is None:
-            return
-        self.log(f"Template option selected: {self.selected_template}")
-        self.dismiss(["template", self.selected_template])
-
-    @on(Button.Pressed, "#select_process_btn")
-    def handle_process_selected(self, event: Button.Pressed) -> None:
-        """Handle the selection of a process option."""
-        if self.selected_process is None:
-            return
-        self.log(f"Process option selected: {self.selected_process}")
-        self.dismiss(["process", self.selected_process])
-
-
-class TechnologyTypesScreen(ModalScreen):
-    """Modal screen to display technology types in Egeria."""
-    BINDINGS = [("q", "dismiss(200)", "Quit"),
-                ("ctl+e", "expand", "Toggle Twisties")]
-
-    CSS_PATH = "my_profile.tcss"
-
-    def __init__(self, ttlist, user_name, user_pwd, user_kpts):
-        """Initialize the TechnologyTypes screen with a list of technology types."""
-        super().__init__()
-        self.tech_type_list = ttlist
-        self.user_name = user_name
-        self.user_password = user_pwd
-        self.karma_points = user_kpts
-        self.tech_type_tree: Tree[str] = Tree(label="Technology Types", id="technology_types_tree")
-        self.tech_type_tree.root.expand()
-        self.tech_type_tree.auto_expand = True
-        self.selected_t_node = None
-        self.selected_t_node_label = None
-        self.node_id = None
-        self.node_status = "expanded"
-        load_app_config()
-        app_config = settings.Environment
-        app_user = settings.User_Profile
-        # config_logging()
-        self.user_name = app_user.user_name or "garygeeke"
-        self.user_password = app_user.user_pwd or "secret"
-
-    def on_mount(self) -> None:
-        self.title = f"User: {self.user_name}"
-        self.sub_title = "Select a technology type"
-
-    def compose(self) -> ComposeResult:
-        """ Compose and display the technology type screen"""
-        self.tech_type_tree.refresh()
-        if self.tech_type_list:
-            self.log(f"Technology types: {self.tech_type_list}, type: {type(self.tech_type_list)}")
-            self.render_tech_type_hierarchy_to_tree(self.tech_type_list, self.tech_type_tree)
-        else:
-            self.tech_type_tree.root.add("No technology types found", expand=True)
-        self.tech_type_tree.refresh()
-        self.log(f"Technology types tree: {self.tech_type_tree}")
-
-        yield Header(show_clock=True)
-        yield ScrollableContainer(
-            Static("Display technology types in Egeria"),
-            self.tech_type_tree,
-            Button("Select", id="select_tech_type_btn"),
-            id="technology_types_table"
-        )
-        yield Footer()
-
-    def action_quit(self) -> None:
-        """ The quit option in the footer has been selected. Dismiss the screen."""
-        self.dismiss("200")
-
-    @on(Button.Pressed, "#select_tech_type_btn")
-    def handle_select_tech_type(self, event: Button.Pressed) -> str|None:
-        """ The select button on the screen has been pressed."""
-        self.log(f"Select button pressed, button: {event.button}")
-        if self.selected_t_node:
-            self.log(f"Selected node: {self.selected_t_node}, label: {self.selected_t_node.label}")
-            self.dismiss(str(self.selected_t_node.label))
-        else:
-            return "No technology type selected"
-
-    @on(Tree.NodeSelected)
-    def handle_tree_node_selected(self, event: Tree.NodeSelected) -> None:
-        """ The user has selected a node in the tree. """
-        self.log(f"Tree node selected, node: {event.node}")
-        self.selected_t_node = event.node
-        self.selected_t_node_label = event.node.label
-        return
-
-    @on(Tree.NodeCollapsed)
-    def handle_tree_node_collapsed(self, event: Tree.NodeCollapsed) -> None:
-        # handle the twisty to close a node in the tree
-        # logger.debug(f"TreeNodeCollapsed: {event.node.id}")
-        self.node_id = str(event.node.id)
-        self.node_status = "collapsed"
-
-    @on(Tree.NodeExpanded)
-    def handle_tree_node_expanded(self, event: Tree.NodeExpanded) -> None:
-        # handle the twisty to open a node in the tree
-        # logger.debug(f"TreeNodeExpanded: {event.node.id}")
-        self.node_id = str(event.node.id)
-        self.node_status = "expanded"
-
-    def action_expand(self):
-        # handle the hot key to expand/collapse all nodes in the tree (see bindings)
-        self.tech_type_tree = self.query_one("#technology_types_tree", Tree)
-        if self.node_status == "collapsed":
-            self.tech_type_tree.root.expand_all()
-        else:
-            self.tech_type_tree.root.collapse_all()
-        return
-
-    def render_tech_type_hierarchy_to_tree(self,
-            data: dict | list[dict],
-            tree_or_node: Tree[str] | TreeNode[str],
-            label_attr: str = "displayName",
-            guid_attr: str = "guid",
-            children_attr: str = "subTypes"
-    ) -> None:
-        """Recursively render a DICT) hierarchy into a Textual Tree.
-
-        Args:
-            data: The data structure from pyegeria.get_tech_type_hierarchy.
-            tree_or_node: The Tree object or a TreeNode to add children to.
-            label_attr: The attribute in the data for the node label.
-            guid_attr: The attribute in the data for the node data (GUID).
-            children_attr: The attribute in the data containing the list of children.
-        """
-        if isinstance(data, list):
-            for item in data:
-                self.render_tech_type_hierarchy_to_tree(item, tree_or_node, label_attr, guid_attr, children_attr)
-            return
-
-        if not isinstance(data, dict):
-            return
-
-        label = str(data.get(label_attr) or data.get("Display Name") or "Unknown")
-        guid = str(data.get(guid_attr) or data.get("GUID") or "")
-
-        # If it's a Tree, we add to root. If it's a TreeNode, we add to it.
-        if isinstance(tree_or_node, Tree):
-            node = tree_or_node.root.add(label, data=guid, expand=True)
-        else:
-            node = tree_or_node.add(label, data=guid, expand=True)
-
-        children = data.get(children_attr)
-        if children and isinstance(children, list):
-            for child in children:
-                self.render_tech_type_hierarchy_to_tree(child, node, label_attr, guid_attr, children_attr)
-
-
-class CreateProfileScreen(ModalScreen[int]):
-    """Modal screen to create a new user profile in Egeria.
-
-    Dismisses with:
-      200 on success
-      4xx on failure
-    """
-
-    BINDINGS = [("q", "dismiss(200)", "Quit")]
-    CSS_PATH = "my_profile.tcss"
-
-    def __init__(self):
-        super().__init__()
-        load_app_config()
-        app_config = settings.Environment
-        app_user = settings.User_Profile
-        # config_logging()
-        self.karma_points = 0
-
-        self.user_name = app_user.user_name or "garygeeke"
-        self.user_password = app_user.user_pwd or "secret"
-        self.view_server = app_config.egeria_view_server
-        self.platform_url = app_config.egeria_platform_url
-        print("Platform:", app_config.egeria_platform_url)
-        print("View Server:", app_config.egeria_view_server)
-
-    def on_mount(self) -> None:
-        self.title = f"User: {self.user_name}, Karma Points: {self.karma_points}"
-        self.sub_title = f"Create Egeria Profile for user: {self.user_name}"
-
-    def compose(self) -> ComposeResult:
-        yield Header(show_clock=True)
-        yield Static("Create a new profile in Egeria")
-        yield ScrollableContainer(
-            Static("Fill in your information below:"),
-            Input(placeholder="Courtesy Title", id="user_title"),
-            Input(placeholder="Given Names", id="user_given_names"),
-            Input(placeholder="Family Name", id="user_family_name"),
-            Input(placeholder="Preferred Name", id="user_preferred_name"),
-            Input(placeholder="Pronouns", id="user_pronouns"),
-            Input(placeholder="Job Title", id="user_job_title"),
-            Input(placeholder="Description", id="user_description"),
-            Input(placeholder="Employee ID", id="user_employee_id"),
-            Input(placeholder="Preferred Language", id="user_preferred_language"),
-            Input(placeholder="Resident Country", id="user_resident_country"),
-            Input(placeholder="Time Zone", id="user_time_zone"),
-            Button("Create Profile", id="create_profile_btn"),
-            id="create_profile_form",
-        )
-
-        yield Footer()
-
-    @on(Button.Pressed, "#create_profile_btn")
-    def create_profile(self) -> int:
-        """Create a new profile in Egeria from data provided in Input fields."""
-
-        self.new_element_request_body:dict = {"class": "NewElementRequestBody",
-                                              "isOwnAnchor": True,
-                                              "properties": {
-                                                  "class": "PersonProperties",
-                                                  "qualifiedName": "Person" + self.query_one("#user_employee_id",
-                                                                                             Input).value +
-                                                                   self.query_one("#user_resident_country",
-                                                                                  Input).value +
-                                                                   self.query_one("#user_given_names", Input).value +
-                                                                   self.query_one("#user_family_name", Input).value,
-                                                  "displayName": self.query_one("#user_preferred_name", Input).value,
-                                                  "courtesyTitle": self.query_one("#user_title", Input).value,
-                                                  "initials": "PAT",
-                                                  "givenNames": self.query_one("#user_given_names", Input).value,
-                                                  "surname": self.query_one("#user_family_name", Input).value,
-                                                  "fullName": self.query_one("#user_preferred_name", Input).value,
-                                                  "pronouns": self.query_one("#user_pronouns", Input).value,
-                                                  "jobTitle": self.query_one("#user_job_title", Input).value,
-                                                  "employeeNumber": self.query_one("#user_employee_id", Input).value,
-                                                  "employeeType": "Full-Time",
-                                                  "preferredLanguage": self.query_one("#user_preferred_language",
-                                                                                      Input).value,
-                                                  "residentCountry": self.query_one("#user_resident_country",
-                                                                                    Input).value,
-                                                  "timeZone": self.query_one("#user_time_zone", Input).value,
-                                                  "description": self.query_one("#user_description", Input).value,
-                                              }
-                                            }
-        try:
-            new_profile_inst = MyProfile(self.app.view_server, self.app.platform_url, self.app.user, self.app.password)
-            new_profile_guid = new_profile_inst.add_my_profile(self.new_element_request_body)
-            self.log(f"Profile created with GUID: {new_profile_guid}")
-            self.dismiss(200)
-            return(200)
-        except PyegeriaException as e:
-            self.log(f"Error creating profile: {e!s} | request={self.new_element_request_body}")
-            self.dismiss(401)
-            return(401)
-
-    def action_quit(self) -> int:
-        self.dismiss(200)
-        return(200)
-
-
-class MyProfileTui(App):
+from CreateProfileScreen import CreateProfileScreen
+from TechnologyTypesScreen import TechnologyTypesScreen
+from TechnologyTypeOptionsScreen import TechnologyTypeOptionsScreen
+from TechnologyTypeTemplatesScreen import TechnologyTypeTemplatesScreen
+from TechnologyTypeProcessesScreen import TechnologyTypeProcessesScreen
+from StatusScreen import StatusScreen
+from ShopForDataScreen import ShopForDataScreen
+from SelectionOverviewScreen import SelectionOverviewScreen
+
+
+class MyProfileApp(App):
     """My Profile App.
 
     Retrieves a user's profile from Egeria and displays current work items.
@@ -1770,7 +687,7 @@ class MyProfileTui(App):
             # then use the pyegeria create_metadata_element_from template to create the metadata element
             # finally display a status screen to confirm the creation of the metadata element
 
-        # make a working version of the full template if it is returned 
+        # make a working version of the full template if it is returned
         if isinstance(result[2], dict):
             self.full_template = result[2]
         # create an empty dict for the returned data
@@ -2199,5 +1116,5 @@ class MyProfileTui(App):
         pass
 
 if __name__ == "__main__":
-    app = MyProfileTui()
+    app = MyProfileApp()
     app.run()
